+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Support for grouped, optional, and repeated symbols.
 - Extended grammar in the `ext` module.
+- Switch to dump the parsed grammar (`--dump-grammar`).
+- Switch to dump the parsed extended grammar (`--dump-ext-grammar`).
 
 ## 0.4.1 - 2018-08-13
 ### Changed